replace generic message lane APIs with chain-specific
