+++ primitives/message-lane/Cargo.toml
@@ -12,7 +12,6 @@ codec = { package = "parity-scale-codec", version = "1.3.1", default-features = 
 # Substrate Dependencies
 
 frame-support = { version = "2.0", default-features = false }
-sp-api = { version = "2.0", default-features = false }
 sp-std = { version = "2.0", default-features = false }
 
 [features]
@@ -20,6 +19,5 @@ default = ["std"]
 std = [
 	"codec/std",
 	"frame-support/std",
-	"sp-api/std",
 	"sp-std/std"
 ]
