+++ primitives/message-lane/src/lib.rs
@@ -24,7 +24,6 @@
 
 use codec::{Decode, Encode};
 use frame_support::RuntimeDebug;
-use sp_api::decl_runtime_apis;
 use sp_std::{collections::vec_deque::VecDeque, prelude::*};
 
 pub mod source_chain;
@@ -123,32 +122,5 @@ impl Default for OutboundLaneData {
 			latest_received_nonce: 0,
 			latest_generated_nonce: 0,
 		}
-	}
-}
-
-decl_runtime_apis! {
-	/// Outbound message lane API.
-	pub trait OutboundLaneApi {
-		/// Returns dispatch weight of all messages in given inclusive range.
-		///
-		/// If some (or all) messages are missing from the storage, they'll also will
-		/// be missing from the resulting vector. The vector is ordered by the nonce.
-		fn messages_dispatch_weight(
-			lane: LaneId,
-			begin: MessageNonce,
-			end: MessageNonce,
-		) -> Vec<(MessageNonce, Weight)>;
-		/// Returns nonce of the latest message, received by bridged chain.
-		fn latest_received_nonce(lane: LaneId) -> MessageNonce;
-		/// Returns nonce of the latest message, generated by given lane.
-		fn latest_generated_nonce(lane: LaneId) -> MessageNonce;
-	}
-
-	/// Inbound message lane API.
-	pub trait InboundLaneApi {
-		/// Returns nonce of the latest message, received by given lane.
-		fn latest_received_nonce(lane: LaneId) -> MessageNonce;
-		/// Nonce of latest message that has been confirmed to the bridged chain.
-		fn latest_confirmed_nonce(lane: LaneId) -> MessageNonce;
 	}
 }
